Finished project functionality and design related logic
