+++ index.html
@@ -16,37 +16,28 @@
         <span></span>
       </div>
     </div>
-
     <div class="app-screen">
       <div class="logo">Weathery</div>
+      <div class="separator"></div>
       <div class="search-bar-wrapper">
-        <form class="input-form">
-          <div class="form-inner-wrapper">
-            <input class="search-input" type="text" name="location" autocomplete="off" placeholder="Location">
-            <button class="submit-button" type="submit"><span></span></button>
+        <div class="search-label">Location</div>
+        <div class="search-bar-wrapper">
+          <form class="input-form">
+            <div class="form-inner-wrapper">
+              <input class="search-input" type="text" name="location" autocomplete="off" placeholder="City name">
+              <button class="submit-button" type="submit"><span></span></button>
+            </div>
+          </form>
+          <span></span>
+          <span></span>
+          <div class="search-bar">
+            <span></span>
           </div>
-        </form>
-
-        <span></span>
-        <span></span>
-        <div class="search-bar">
-          <span></span>
         </div>
       </div>
-
-      <div class="pin"></div>
-      <div class="location"></div>
-      <div class="date-and-time"></div>
-      <div class="current-temperature">
-        <img class="icon" alt="Weather icon"></img>
-        <div class="degrees">
-          <span></span>
-          <span>&#176;</span>
-        </div>
+      <div class="search-result-wrapper">
+        <div class="search-result-inner-wrapper"></div>
       </div>
-      <div class="temperature-info">
-      </div>
-      <div class="description"></div>
       <div class="unit-toggler-wrapper">
         <div>&#8451;</div>
         <div class="unit-toggler">
@@ -54,10 +45,8 @@
         </div>
         <div>&#8457;</div>
       </div>
+      <div class="separator"></div>
     </div>
-
   </div>  
-  <!-- <img class="giphy" src="http://i.stack.imgur.com/SBv4T.gif" alt="this slowpoke moves"  width=200/> -->
-  <!-- <img class="giphy" src="http://i.stack.imgur.com/SBv4T.gif" alt="this slowpoke moves"  width=200/> -->
 </body>
 </html>
