+++ src/css/index.scss
@@ -21,6 +21,7 @@ $animation-duration: 2s;
   max-height: 1200px;
   width: 80%;
   height: 80%;
+  min-height: 400px;
 
   box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
   background: inherit;
@@ -41,21 +42,40 @@ $animation-duration: 2s;
 // Overall
 
 .logo {
-  position: absolute;
-  top: 30px;
-  font-size: 40px;
-  color: white;
+  position: relative;
+  margin-top: 30px;
+  font-size:calc(20px + 0.8vw);
+  color: $color;
+}
+
+.separator {
+  flex-grow: 1;
+}
+
+.search-result-wrapper {
+  height: 0;
+  opacity: 0;
+  transition: height 1s ease, opacity 0.4s ease 0.8s; 
+  will-change: transform, opacity;
+}
+
+.search-result-inner-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
 
 .location {
   position: relative;
   margin-top: 10px;
-  color: white;
+  font-size:calc(12px + 0.6vw);
+  color: $color;
 }
 
 .date-and-time {
   position: relative;
   margin-top: 10px;
+  font-size:calc(10px + 0.5vw);
   color: rgba(255, 255, 255, 0.8);
 }
 
@@ -79,12 +99,12 @@ $animation-duration: 2s;
 
   .degrees {
     position: relative;
-    font-size: 60px;
-    color: white;
+    font-size:calc(30px + 1vw);
+    color: $color;
 
     span:nth-of-type(2) {
       position: absolute;
-      font-size: 30px;
+      font-size:calc(15px + 1vw);
       top: 4px;
     }
   }
@@ -92,7 +112,8 @@ $animation-duration: 2s;
 
 .temperature-info {
   position: relative;
-  color: white;
+  font-size:calc(10px + 0.5vw);
+  color: $color;
 
   span:nth-of-type(1) {
     height: 100%;
@@ -103,8 +124,8 @@ $animation-duration: 2s;
 .description {
   position: relative;
   margin-top: 30px;
-  font-size: 24px;
-  color: white;
+  font-size:calc(12px + 0.8vw);
+  color: $color;
 }
 
 // Unit toggler
@@ -112,8 +133,8 @@ $animation-duration: 2s;
 .unit-toggler-wrapper {
   display: flex;
   align-items: center;
-  position: absolute;
-  bottom: 30px;
+  position: relative;
+  margin: 20px 0 30px 0;
 }
 
 .unit-toggler {
@@ -131,7 +152,7 @@ $animation-duration: 2s;
     width: 16px;
     height: 16px;
     border-radius: 50%;
-    background-color: white;
+    background-color: $color;
     transition: transform 0.3s ease;
     will-change: transform;
 
@@ -173,19 +194,19 @@ $animation-duration: 2s;
     width: 20px;
     height: 20px;
     z-index: 99999;
-    color: white;
+    color: $color;
 
     span:nth-of-type(1) {
       position: absolute;
       width: 100%;
-      border-top: 1px solid white;
+      border-top: 1px solid $color;
       transform: rotate(-45deg);
     }
     
     span:nth-of-type(2) {
       position: absolute;
       width: 100%;
-      border-top: 1px solid white;
+      border-top: 1px solid $color;
       transform: rotate(45deg);
     }
 
@@ -216,7 +237,7 @@ $animation-duration: 2s;
 .pin {
   position: relative;
   width: 30px;
-  height: 30px;
+  padding-bottom: 30px;
   margin-top: 20px;
   background-image: url('../images/pin-thin.svg');
   background-position: center;
@@ -225,6 +246,50 @@ $animation-duration: 2s;
 }
 
 // Search bar
+
+.search-bar-wrapper {
+  display: flex;
+  align-items: center;
+  position: relative;
+
+  > span:nth-of-type(1) {
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    height: 60%;
+    border-left: 1px solid $color;
+    opacity: 0;
+    transform-origin: bottom;
+    animation: slide-rotate-in time(0.25) linear forwards;
+    transition: opacity 0s linear time(0.375);
+    will-change: transform, opacity;
+  }
+  
+  > span:nth-of-type(2) {
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    width: $initial-width;
+    border-bottom: 1px solid $color;
+    opacity: 0;
+    transition: opacity 0s linear time(0.375), transform time(0.125) linear;
+    will-change: transform, opacity;
+  }
+
+  &:hover {
+    cursor: pointer;
+  }
+}
+
+.search-label {
+  position: relative;
+  width: 80px;
+  margin-right: 8px;
+  font-size: 20px;
+  opacity: 1;
+  transition: opacity 1s ease, width time(0.125) linear time(0.65);
+  will-change: opacity, width;
+}
 
 .search-bar-wrapper form {
   position: absolute;
@@ -239,11 +304,11 @@ $animation-duration: 2s;
     border: none;
     outline: none;
     background: transparent;
-    color: white;
+    color: $color;
 
     &::placeholder { /* Chrome, Firefox, Opera, Safari 10.1+ */
       padding-left: 2px;
-      color: white;
+      color: $color;
       opacity: 1; /* Firefox */
     }
   }
@@ -260,7 +325,7 @@ $animation-duration: 2s;
     height: 14px;
     padding: 0;
     border-radius: 50%;
-    border: 1px solid white;
+    border: 1px solid $color;
     outline: none;
     background: transparent;
     transform: translateY(-50%);
@@ -270,7 +335,7 @@ $animation-duration: 2s;
       top: 100% * 0.8535;
       left: 100% * 0.8535;
       width: 10px;
-      border-top: 1px solid white;
+      border-top: 1px solid $color;
       transform-origin: left;
       transform: rotate(45deg);  
     }
@@ -278,36 +343,6 @@ $animation-duration: 2s;
     &:hover {
       cursor: pointer;
     }
-  }
-}
-
-.search-bar-wrapper {
-  position: relative;
-
-  > span:nth-of-type(1) {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    height: 60%;
-    border-left: 1px solid $color;
-    opacity: 0;
-    transform-origin: bottom;
-    animation: slide-rotate-in time(0.25) linear forwards;
-    transition: opacity 0s linear time(0.375);
-  }
-  
-  > span:nth-of-type(2) {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 100%;
-    border-bottom: 1px solid $color;
-    opacity: 0;
-    transition: opacity 0s linear time(0.375), transform time(0.125) linear;
-  }
-
-  &:hover {
-    cursor: pointer;
   }
 }
 
@@ -319,50 +354,62 @@ $animation-duration: 2s;
   border-radius: 50%;
   opacity: 1;
   transition: border-radius time(0.25) linear time(0.375), opacity time(0.25) linear time(0.25);
+  will-change: border-radius, opacity;
   
   span:nth-of-type(1) {
     position: absolute;
-    top: 100% * 0.8535;
-    left: 100% * 0.8535;
+    top: 24px;
+    left: 24px;
     width: $initial-width * 0.7;
     border-top: 1px solid $color;
     transform-origin: left;
     transform: rotate(45deg);  
     transition: transform time(0.25) linear time(0.55);
+    will-change: transform;
   }
 }
 
 .search-bar-wrapper.active {
+  .search-label {
+    opacity: 0;
+    width: 0;
+  }
+
   > span:nth-of-type(1) {
     opacity: 1;
     transform-origin: bottom;
     animation: slide-rotate-out time(0.25) linear time(0.525) forwards;
     transition: opacity 0s linear time(0.375);
+    will-change: opacity;
   }
   
   > span:nth-of-type(2) {
     opacity: 1;
     transform: scaleX($full-width / $initial-width);
     transition: opacity 0s linear time(0.375), transform time(0.125) linear time(0.65);
+    will-change: opacity, transform;
   }
-
+  
   .search-bar {
     border-radius: 0;
     opacity: 0;
     transition: border-radius time(0.25) linear time(0.2), opacity time(0.175) linear time(0.35);
+    will-change: border-radius, opacity;
     
     span:nth-of-type(1) {
       opacity: 0;
       transform: rotate(45deg) scaleX(0);
       transition: transform time(0.25) linear, opacity 0s time(0.2);
+      will-change: transform, opacity;
     }
   }
-
+  
   form {
     opacity: 1;
-    transform: translate(calc(-50% + calc(#{$initial-width} * 0.5)), calc(-50% + calc(#{$initial-height} * 0.5)));
-    transition: transform 0s time(0.8), opacity time(0.25) ease time(0.8);
     z-index: 333;
+    transform: translate(calc(-50% + calc(#{$initial-width} * 0.5)), calc(-50% + 10px));
+    transition: transform 0s time(0.8), opacity time(0.25) ease time(0.8);
+    will-change: transform, opacity;
   }
 }
