+++ src/css/style.scss
@@ -1,40 +1,4 @@
 @font-face {
-  font-family: 'Lobster';
-  src: url('../fonts/Lobster-Regular.eot');
-  src: url('../fonts/Lobster-Regular.eot?#iefix') format('embedded-opentype'),
-      url('../fonts/Lobster-Regular.woff2') format('woff2'),
-      url('../fonts/Lobster-Regular.woff') format('woff'),
-      url('../fonts/Lobster-Regular.ttf') format('truetype'),
-      url('../fonts/Lobster-Regular.svg#Lobster-Regular') format('svg');
-  font-weight: normal;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'Quicksand';
-  src: url('../fonts/Quicksand-Bold.eot');
-  src: url('../fonts/Quicksand-Bold.eot?#iefix') format('embedded-opentype'),
-      url('../fonts/Quicksand-Bold.woff2') format('woff2'),
-      url('../fonts/Quicksand-Bold.woff') format('woff'),
-      url('../fonts/Quicksand-Bold.ttf') format('truetype'),
-      url('../fonts/Quicksand-Bold.svg#Quicksand-Bold') format('svg');
-  font-weight: bold;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'Quicksand';
-  src: url('../fonts/Quicksand-SemiBold.eot');
-  src: url('../fonts/Quicksand-SemiBold.eot?#iefix') format('embedded-opentype'),
-      url('../fonts/Quicksand-SemiBold.woff2') format('woff2'),
-      url('../fonts/Quicksand-SemiBold.woff') format('woff'),
-      url('../fonts/Quicksand-SemiBold.ttf') format('truetype'),
-      url('../fonts/Quicksand-SemiBold.svg#Quicksand-SemiBold') format('svg');
-  font-weight: 600;
-  font-style: normal;
-}
-
-@font-face {
   font-family: 'Quicksand';
   src: url('../fonts/Quicksand-Regular.eot');
   src: url('../fonts/Quicksand-Regular.eot?#iefix') format('embedded-opentype'),
@@ -43,30 +7,6 @@
       url('../fonts/Quicksand-Regular.ttf') format('truetype'),
       url('../fonts/Quicksand-Regular.svg#Quicksand-Regular') format('svg');
   font-weight: normal;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'Quicksand';
-  src: url('../fonts/Quicksand-Light.eot');
-  src: url('../fonts/Quicksand-Light.eot?#iefix') format('embedded-opentype'),
-      url('../fonts/Quicksand-Light.woff2') format('woff2'),
-      url('../fonts/Quicksand-Light.woff') format('woff'),
-      url('../fonts/Quicksand-Light.ttf') format('truetype'),
-      url('../fonts/Quicksand-Light.svg#Quicksand-Light') format('svg');
-  font-weight: 300;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'Quicksand';
-  src: url('../fonts/Quicksand-Medium.eot');
-  src: url('../fonts/Quicksand-Medium.eot?#iefix') format('embedded-opentype'),
-      url('../fonts/Quicksand-Medium.woff2') format('woff2'),
-      url('../fonts/Quicksand-Medium.woff') format('woff'),
-      url('../fonts/Quicksand-Medium.ttf') format('truetype'),
-      url('../fonts/Quicksand-Medium.svg#Quicksand-Medium') format('svg');
-  font-weight: 500;
   font-style: normal;
 }
 
@@ -80,7 +20,7 @@
   height: 100vh;
   font-family: 'Quicksand', sans-serif;
   color: white;
-  background-image: url('../images/background-narrow.jpg');
+  background-image: url('../images/background.jpg');
   background-attachment: fixed;
   background-position: center;
   background-repeat: no-repeat;
