+++ src/javascript/utility-functions.js
@@ -1,0 +1,5 @@
+const capitalizeFirstLetter = (string) => {
+  return string.charAt(0).toUpperCase() + string.slice(1);
+}
+
+export { capitalizeFirstLetter };
